Make CancelReferee service generic to enable reuse
